Add test-specific attribute to name input element
Lesson 4

The simplest method for locating the text input element to test is to
search for the input type directly.

However, in the future, more input elements could be added, or the input
type could change. To make our location method more specific, we can
assign a test-specific attribute to the element.