Clear Jest mocks before each test for lesson 5
The second test will fail because it sees two API calls. The mock from
the first test has not been cleared out by Jest. We need to tell Jest to
clearAllMocks() by adding a method to the beginning of the test file
MessageDisplay.spec.js.

Note that `clearMocks: true` can also be specified in the Jest config,
either in package.json or jest.config.js.

https://jestjs.io/docs/en/configuration#clearmocks-boolean